Chapter 5: review comments
(cherry picked from commit 76e01cae83341b728f5f003bbbd34fe67ffcb9d1)

+++ Examples/Modules/Chapter 05/Ex5_03B.cpp
@@ -3,8 +3,8 @@ import std;
 
 int main()
 {
-  const size_t values_per_line {3}; // Outputs per line
-  size_t values_current_line {};    // Number of outputs on current line
+  const unsigned values_per_line {3}; // Outputs per line
+  unsigned values_current_line {};    // Number of outputs on current line
   for (double radius {0.2}; radius < 3.0 + 0.001; radius += 0.2)
   {
     const auto area{ std::numbers::pi * radius * radius };
